Fix a bug in handling supplementary alignments
SAM format specification does not impose an order of alignment lines (i.e. supplementary alignments might appear before or after the representative alignment). Here in code, the first alignment is taken as "primary", all the following as "supplementary", hence the flag is_supplementary should be ignored. The decision regarding which alignment is representative is usually arbitrary.