Make hicExport write to gzipped text files in Python 3.
`gzip.open` by default opens file in binary mode (unlike the regular `open`).
In Python 3, an attempt to write a string to a binary-open file fails with `TypeError: memoryview: a bytes-like object is required, not 'str'`.
To prevent this error, the file should be open in text mode ("wt" instead of "w") in the first place.
See https://stackoverflow.com/questions/41050881/write-to-gzipped-fasta-file-in-python-3-5